+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -3,6 +3,7 @@ import { useEffect, Suspense, useSyncExternalStore } from 'react';
 import DocumentTitleWrapper from './DocumentTitleWrapper';
 import PageLoading from './PageLoading';
 import { useEscapeKeyStroke } from './hooks/useEscapeKeyStroke';
+import { useForceLogout } from './hooks/useForceLogout';
 import { useGoogleTagManager } from './hooks/useGoogleTagManager';
 import { useMessageLinkClicks } from './hooks/useMessageLinkClicks';
 import { useOTRMessaging } from './hooks/useOTRMessaging';
@@ -11,7 +12,6 @@ import { useStoreCookiesOnLogin } from './hooks/useStoreCookiesOnLogin';
 import { useUpdateVideoConfUser } from './hooks/useUpdateVideoConfUser';
 import { useWebRTC } from './hooks/useWebRTC';
 import { useWordPressOAuth } from './hooks/useWordPressOAuth';
-import { useAnalytics } from '../../../app/analytics/client/loadScript';
 import { useCorsSSLConfig } from '../../../app/cors/client/useCorsSSLConfig';
 import { useDolphin } from '../../../app/dolphin/client/hooks/useDolphin';
 import { useDrupal } from '../../../app/drupal/client/hooks/useDrupal';
@@ -22,7 +22,9 @@ import { useLivechatEnterprise } from '../../../app/livechat-enterprise/hooks/us
 import { useNextcloud } from '../../../app/nextcloud/client/useNextcloud';
 import { useTokenPassAuth } from '../../../app/tokenpass/client/hooks/useTokenPassAuth';
 import { useNotificationPermission } from '../../hooks/notification/useNotificationPermission';
+import { useNotificationUserCalendar } from '../../hooks/notification/useNotificationUserCalendar';
 import { useNotifyUser } from '../../hooks/notification/useNotifyUser';
+import { useAnalytics } from '../../hooks/useAnalytics';
 import { useAnalyticsEventTracking } from '../../hooks/useAnalyticsEventTracking';
 import { useAutoupdate } from '../../hooks/useAutoupdate';
 import { useLoadRoomForAllowedAnonymousRead } from '../../hooks/useLoadRoomForAllowedAnonymousRead';
@@ -65,6 +67,8 @@ const AppLayout = () => {
 	useWebRTC();
 	useStoreCookiesOnLogin();
 	useAutoupdate();
+	useForceLogout();
+	useNotificationUserCalendar();
 
 	const layout = useSyncExternalStore(appLayout.subscribe, appLayout.getSnapshot);